Removed WordFilterSO and created wordfilter encoded txt file
