--- UnityProject/Assets/Resources/ScriptableObjects/SOs singletons/WordFilterSO.asset
@@ -1,47 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: af0b3d3b8c59f1d4593b24402d05c42a, type: 3}
-  m_Name: WordFilterSO
-  m_EditorClassIdentifier: 
-  FilterList:
-  - TargetWord: nigga
-    ReplaceWithWord: proud african friend
-  - TargetWord: n1gga
-    ReplaceWithWord: proud african friend
-  - TargetWord: n1gg4
-    ReplaceWithWord: proud african friend
-  - TargetWord: n i g g a
-    ReplaceWithWord: proud african friend
-  - TargetWord: hitler
-    ReplaceWithWord: fopdoodle
-  - TargetWord: faggot
-    ReplaceWithWord: bundle of sticks
-  - TargetWord: fagget
-    ReplaceWithWord: bundle of sticks
-  - TargetWord: faggit
-    ReplaceWithWord: bundle of sticks
-  - TargetWord: fag
-    ReplaceWithWord: a stick
-  - TargetWord: f a g
-    ReplaceWithWord: a stick
-  - TargetWord: f a g g o t
-    ReplaceWithWord: bundle of sticks
-  - TargetWord: nigger
-    ReplaceWithWord: proud african friend
-  - TargetWord: niggers
-    ReplaceWithWord: proud african friends
-  - TargetWord: chink
-    ReplaceWithWord: proud asian friend
-  - TargetWord: chinks
-    ReplaceWithWord: proud asian friends
-  - TargetWord: fagg ot
-    ReplaceWithWord: bundle of sticks
