--- UnityProject/Assets/Scripts/ScriptableObject/WordFilterSO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: af0b3d3b8c59f1d4593b24402d05c42a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
