+++ UnityProject/Assets/StreamingAssets/wordfilter.txt
@@ -1,0 +1,1 @@
+eyJGaWx0ZXJMaXN0IjpbeyJUYXJnZXRXb3JkIjoibmlnZ2EiLCJSZXBsYWNlV2l0aFdvcmQiOiJwcm91ZCBhZnJpY2FuIGZyaWVuZCJ9LHsiVGFyZ2V0V29yZCI6Im4xZ2dhIiwiUmVwbGFjZVdpdGhXb3JkIjoicHJvdWQgYWZyaWNhbiBmcmllbmQifSx7IlRhcmdldFdvcmQiOiJuMWdnNCIsIlJlcGxhY2VXaXRoV29yZCI6InByb3VkIGFmcmljYW4gZnJpZW5kIn0seyJUYXJnZXRXb3JkIjoibiBpIGcgZyBhIiwiUmVwbGFjZVdpdGhXb3JkIjoicHJvdWQgYWZyaWNhbiBmcmllbmQifSx7IlRhcmdldFdvcmQiOiJoaXRsZXIiLCJSZXBsYWNlV2l0aFdvcmQiOiJmb3Bkb29kbGUifSx7IlRhcmdldFdvcmQiOiJmYWdnb3QiLCJSZXBsYWNlV2l0aFdvcmQiOiJidW5kbGUgb2Ygc3RpY2tzIn0seyJUYXJnZXRXb3JkIjoiZmFnZ2V0IiwiUmVwbGFjZVdpdGhXb3JkIjoiYnVuZGxlIG9mIHN0aWNrcyJ9LHsiVGFyZ2V0V29yZCI6ImZhZ2dpdCIsIlJlcGxhY2VXaXRoV29yZCI6ImJ1bmRsZSBvZiBzdGlja3MifSx7IlRhcmdldFdvcmQiOiJmYWciLCJSZXBsYWNlV2l0aFdvcmQiOiJhIHN0aWNrIn0seyJUYXJnZXRXb3JkIjoiZiBhIGciLCJSZXBsYWNlV2l0aFdvcmQiOiJhIHN0aWNrIn0seyJUYXJnZXRXb3JkIjoiZiBhIGcgZyBvIHQiLCJSZXBsYWNlV2l0aFdvcmQiOiJidW5kbGUgb2Ygc3RpY2tzIn0seyJUYXJnZXRXb3JkIjoibmlnZ2VyIiwiUmVwbGFjZVdpdGhXb3JkIjoicHJvdWQgYWZyaWNhbiBmcmllbmQifSx7IlRhcmdldFdvcmQiOiJuaWdnZXJzIiwiUmVwbGFjZVdpdGhXb3JkIjoicHJvdWQgYWZyaWNhbiBmcmllbmRzIn0seyJUYXJnZXRXb3JkIjoiY2hpbmsiLCJSZXBsYWNlV2l0aFdvcmQiOiJwcm91ZCBhc2lhbiBmcmllbmQifSx7IlRhcmdldFdvcmQiOiJjaGlua3MiLCJSZXBsYWNlV2l0aFdvcmQiOiJwcm91ZCBhc2lhbiBmcmllbmRzIn0seyJUYXJnZXRXb3JkIjoiZmFnZyBvdCIsIlJlcGxhY2VXaXRoV29yZCI6ImJ1bmRsZSBvZiBzdGlja3MifV19
