+++ UnityProject/Assets/StreamingAssets/wordfilter.txt.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 8b6c6862a9d89a641816a86601249bf0
-NativeFormatImporter:
+guid: 22998ea4ca00ac049a32fbb11c32d894
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
